(refactor): reorganized directories
