+++ .gitignore
@@ -1,3 +1,4 @@
-.vsocde/*
-vscode/
+.vscode/*
 .vscode/
+.obj/*
+.obj/
