+++ content/Algorithms/BRID/BRID.tpp
@@ -1,0 +1,7 @@
+#include "Algorithms/BRID/BRID.h"
+
+template<typename T>
+std::vector<T> BRID<T>::execute(DataSet<T>& O, std::vector<Cluster<T>>& C) {
+    // TODO: Implementar el algoritmo BRID
+    return std::vector<T>();
+}
