+++ content/Algorithms/MMR/MMR.tpp
@@ -1,0 +1,7 @@
+#include "Algorithms/MMR/MMR.h"
+
+template<typename T>
+std::vector<T> MMR<T>::execute(DataSet<T>& O, std::vector<Cluster<T>>& C) {
+    // TODO: Implementar el algoritmo MMR
+    return std::vector<T>();
+}
