+++ content/Algorithms/Motley/Motley.tpp
@@ -1,0 +1,7 @@
+#include "Algorithms/Motley/Motley.h"
+
+template<typename T>
+std::vector<T> Motley<T>::execute(DataSet<T>& O, std::vector<Cluster<T>>& C) {
+    // TODO: Implementar el algoritmo Motley
+    return std::vector<T>();
+}
