+++ content/Algorithms/Swap/Swap.h
@@ -1,7 +1,7 @@
 #ifndef SWAP_H
 #define SWAP_H
 
-#include "Algorithm.h"
+#include "Algorithms/Algorithm.h"
 
 template<typename T>
 class Swap : public Algorithm<T> {
@@ -14,15 +14,10 @@ public:
          float lambda = 0.5f)
         : Algorithm<T>(sim, div), lambda(lambda) {}
 
-    std::vector<T> execute(DataSet<T>& O, std::vector<Cluster<T>>& C) override {
-        // TODO: Implementar el algoritmo Swap
-        return std::vector<T>();
-    }
-
-    float func() const {
-        // TODO: Implementar función objetivo
-        return 0.0f;
-    }
+    std::vector<T> execute(DataSet<T>& O, std::vector<Cluster<T>>& C) override;
+    float func() const;
 };
+
+#include "Algorithms/Swap/Swap.tpp"
 
 #endif
