+++ content/Algorithms/Swap/Swap.tpp
@@ -1,0 +1,13 @@
+#include "Algorithms/Swap/Swap.h"
+
+template<typename T>
+float Swap<T>::func() const {
+    // TODO: Implementar función objetivo
+    return 0.0f;
+}
+
+template<typename T>
+std::vector<T> Swap<T>::execute(DataSet<T>& O, std::vector<Cluster<T>>& C) {
+    // TODO: Implementar el algoritmo Swap
+    return std::vector<T>();
+}
