+++ content/DataStructures/Data/Data.cpp
@@ -1,4 +1,4 @@
-#include "Data.h"
+#include "DataStructures/Data/Data.h"
 
 template<typename T>
 Data<T>::Data(const T& content) : content(content) {}
