+++ content/DataStructures/Data/Data.h
@@ -1,7 +1,7 @@
 #ifndef DATA_H
 #define DATA_H
 
-#include "Point.h"
+#include "DataStructures/Point/Point.h"
 
 template<typename T>
 class Data {
