+++ content/DataStructures/DataSet/DataSet.cpp
@@ -1,4 +1,4 @@
-#include "DataSet.h"
+#include "DataStructures/DataSet/DataSet.h"
 
 template<typename T>
 T& DataSet<T>::getData(int i) {
